src/requirements.txt
fastapi[standard]
uvicorn

motor
pymongo
pydantic[email]
pydantic-settings
python-dotenv
python-jose[cryptography]

bcrypt==4.0.1
passlib[bcrypt]

pytest
pytest-asyncio
httpx

langchain
langgraph
langchain-core
langchain-google-genai
duckduckgo-search
langchain-community

chromadb
langchain-chroma
